JCR-1593
What is the problem to fix?
    Using webdav with https, it's impossible to rename or move a file or directory. The applied Architecture is an Apache in front which manages the https and AJP connectors that transfer requests to the server JBoss

Problem analysis
This error is caused by the following line of code in WebDavServiceImpl.java:

destinationHeader = serverURI + escapePath(destinationHeader.substring(serverURI.length()));

The output of this line is https://localhost:443/rest/private/jcr/repositorylaboration/ so the workspace collaboration is not found. The consideration of the port 443 is badly done (By using Apache, we don't specify the port).
The path obtained by this line of code:

destinationHeader = TextUtil.unescape(destinationHeader, '%'); is correct (without the specification of the port 443)

There is no problem with http.

How is the problem fixed?
    Change the algorithm of parsing destination URI and base URI. Now we use java.net.URI which itself does all the parsing.