JCR-1597
What is the problem to fix?
There are some problems when using webdav on Windows 7. We can't open a document (MS Office 2010) or the document is opened as read only (Open Office)

How is the problem fixed?
Lock tokens parsing fixed, to parse tokens not surrounded by '<' and '>' characters. To open documents via basic authentication on Windows 7 with MS Office 2010 we also need to edit registry.