JCR-1611
What is the problem to fix?
We need to backport the fix of the jira EXOJCR-1234 to JCR 1.12.x.

How is the problem fixed?
Change the calculation way of last order number during node adding. The new order number will be as real last order number plus 1. 
In previous impl it was as children nodes count plus 1.