Fix random fail in Unit tests (#66)
This random fail is due to the fact that Unit Test methods execution order could be different. In fact, when a 'node creation' only is added in queue, the count of queue elements equals to the last executed operation. When a 'delete or update' operation is made, the size of queue should be less than the last executed queue operation id because in case of 'delete' or 'update', some queue elements are purged, thus the count will be always less than the last ID of the queue.