TASK-38913 Avoid throwing an exception when cleaning SessionProvider twice
Returning a boolean is more adequate here since it shouldn't fail a whole user query when trying to just make a cleanup of resources. Throwing an exception should be made to prevent incoherent logic to be made, to be used also to ensure data consistency... But when cleaning up, just returning a boolean is sufficient to let upper layer know if the SessionProvider was already closed or not.

In addition in this PR, useless and long running Unit Test for deprecated Backup/Restore procedure are evicted.