Task-55630: Element couldn't be deleted (send to trash)
Problem: sometimes, a message error appears when trying to delete a WebContent from a specific site or trash.
Fix: this error owing to getting empty properties of nodeData in phase move to trash or delete from trash by calling the function visitChildProperties (NodeData node).In this fix, I catch this error and continue the process of moving or deleting permanently from the trash.