fix: Change default charset and collation of NAME field of JCR_SITEM table from latin1 to utf8mb4 - EXO-60718 (#286)
Prior to this change, when starting an empty server with Mysql DB and utf8mb4_unicode_ci collation for JCR schema, adding a jcr content whose name contains some
special chars like emojies generate a problem of illegal mix of colllations on NAME field of JCR_SITEM table. After this change, to fix the mix of collations exception, we ensure to create the NAME field of JCR_SITEM table with the same JCR schema collation.