fix: Error when starting Server, class javax.ws.core.Link not found - EXO-62860 - Meeds-io/MIPs#42
Before this fix, and after the merge of MIP-42, the server does not start because the class javax.ws.core.Link was not found
It is due to new jars added in jcr via dependencies on Tika which bring dependencies on org.apache.cxf which do not need.
This commit exclude the problematic wars.