fix: Fix Huge Memory consumption in WebDav Access - EXO-68314 (#308)
Prior to this change, when accessing webdav and getting a property or check it, the jcr:data content is parsed as string and added in properties list. This change will make the property value lazily loaded only when really needed and not when building the Property object.

(cherry picked from commit 6132294af960c3139085fd0abc047462c98b6d9e)